+++ Coursework1/Matrix.hpp
@@ -4,7 +4,7 @@ class Matrix
 {
 public:
     int dim;
-    vector<vector<float>> mat;
+    vector<vector<double>> mat;
 
     // constructor that takes input from the CLI
     // Matrix(int x)
@@ -12,8 +12,8 @@ public:
     //     dim = x;
     //     for (int i = 0; i < dim; ++i)
     //     {
-    //         float input;
-    //         vector<float> tempVec;
+    //         double input;
+    //         vector<double> tempVec;
     //         for (int j = 0; j < dim; ++j)
     //         {
     //             cout << "Enter matrix element " << i + 1 << " , " << j + 1 << ": ";
@@ -26,8 +26,8 @@ public:
     //     this->dim > dim;
     // }
 
-    // contructor that takes a vector of vectors of floats
-    Matrix(vector<vector<float>> in_mat)
+    // contructor that takes a vector of vectors of doubles
+    Matrix(vector<vector<double>> in_mat)
     {
         mat = in_mat;
         dim = mat.size();
@@ -36,7 +36,7 @@ public:
     }
 
 
-    bool augment_matrix(vector<float> vec)
+    bool augment_matrix(vector<double> vec)
     {
         if (vec.size() != dim)
         {
@@ -53,15 +53,15 @@ public:
         return true;
     }
 
-    vector<float> get_row(int x)
+    vector<double> get_row(int x)
     {
-        vector<float> temp = mat[x];
+        vector<double> temp = mat[x];
         return temp;
     }
 
-    vector<float> get_column(int x)
+    vector<double> get_column(int x)
     {
-        vector<float> temp;
+        vector<double> temp;
         for (int i = 0; i < dim; ++i)
         {
             temp.push_back(mat[i][x]);
@@ -71,15 +71,15 @@ public:
 
     void swap_rows(int row1, int row2)
     {
-        vector<float> temp = get_row(row1);
+        vector<double> temp = get_row(row1);
         mat[row1] = get_row(row2);
         mat[row2] = temp;
     }
 
-    float get_abs_max(vector<float> vec)
+    double get_abs_max(vector<double> vec)
     {
-        float max;
-        float temp;
+        double max;
+        double temp;
 
         max = abs(vec[0]);
 
@@ -100,8 +100,8 @@ public:
     {
         for (int i = 0; i < dim; ++i)
         {
-            vector<float> temp = get_row(i);
-            float max = get_abs_max(temp);
+            vector<double> temp = get_row(i);
+            double max = get_abs_max(temp);
             for (int j = 0; j < temp.size(); ++j)
             {
                 mat[i][j] = temp[j] / max;
@@ -109,7 +109,7 @@ public:
         }
     }
 
-    void matrix_row_operation(float coeff, int pivot_row_ix, int target_row_ix)
+    void matrix_row_operation(double coeff, int pivot_row_ix, int target_row_ix)
     {
         for (int i = 0; i < dim + 1; ++i)
         {
@@ -123,7 +123,7 @@ public:
         {
             if (mat[i][i] != 1 && mat[i][i] != 0)
             {
-                float coeff = 1 / mat[i][i];
+                double coeff = 1 / mat[i][i];
                 for (int j = 0; j < dim + 1; ++j)
                 {
                     if (mat[i][j] != 0)
@@ -152,7 +152,7 @@ public:
 
     bool flag_ill_conditioned()
     {
-        vector<float> temp;
+        vector<double> temp;
         for (int i = 0; i < dim - 1; ++i)
         {
             for (int j = i + 1; j < dim; ++j)
@@ -175,7 +175,7 @@ public:
     {
         for (int i = 0; i < dim; i++)
         {
-            float sum = 0;
+            double sum = 0;
             for (int j = 0; j < dim; j++)
                 sum += abs(mat[i][j]);
